src/ml/service.py
from fastapi import FastAPI, File, UploadFile
from pydantic import BaseModel
from typing import List, Dict, Any
import torch
import torch.nn as nn
import torchvision.transforms as transforms
from PIL import Image
import io
import json

app = FastAPI()

class DesignInput(BaseModel):
    design_data: Dict[str, Any]
    settings: Dict[str, Any] = {}

class GeneratedCode(BaseModel):
    html: str
    css: str
    assets: List[Dict[str, str]]

# Basic CNN encoder for processing design images
class DesignEncoder(nn.Module):
    def __init__(self):
        super().__init__()
        self.cnn = nn.Sequential(
            nn.Conv2d(3, 64, kernel_size=3, padding=1),
            nn.ReLU(),
            nn.MaxPool2d(2),
            nn.Conv2d(64, 128, kernel_size=3, padding=1),
            nn.ReLU(),
            nn.MaxPool2d(2),
            nn.Conv2d(128, 256, kernel_size=3, padding=1),
            nn.ReLU(),
            nn.AdaptiveAvgPool2d((1, 1))
        )
        
    def forward(self, x):
        return self.cnn(x).squeeze()

# Initialize model (in production, load pre-trained weights)
model = DesignEncoder()
transform = transforms.Compose([
    transforms.Resize((224, 224)),
    transforms.ToTensor(),
    transforms.Normalize(mean=[0.485, 0.456, 0.406], 
                       std=[0.229, 0.224, 0.225])
])

def process_design(design_data: Dict[str, Any]) -> GeneratedCode:
    """
    Process the design data and generate HTML/CSS code.
    This is a placeholder implementation - you'll need to implement
    the actual ML model logic.
    """
    # Placeholder implementation
    return GeneratedCode(
        html="<div class='container'>\n  <!-- Generated content will go here -->\n</div>",
        css=".container {\n  /* Generated styles will go here */\n}",
        assets=[]
    )

@app.post("/convert", response_model=GeneratedCode)
async def convert_design(design_input: DesignInput):
    """
    Convert a Figma design into HTML/CSS code
    """
    try:
        # Process the design data
        generated_code = process_design(design_input.design_data)
        return generated_code
    except Exception as e:
        return {"error": str(e)}

@app.post("/process-image")
async def process_image(file: UploadFile = File(...)):
    """
    Process a single design image and return features
    """
    try:
        # Read and process the image
        contents = await file.read()
        image = Image.open(io.BytesIO(contents)).convert('RGB')
        
        # Transform the image
        img_tensor = transform(image).unsqueeze(0)
        
        # Get features (in production, this would feed into a more complex pipeline)
        with torch.no_grad():
            features = model(img_tensor)
        
        return {"features": features.tolist()}
    except Exception as e:
        return {"error": str(e)}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
