requirements.txt
fastapi>=0.95.0,<1.0.0
uvicorn>=0.20.0,<1.0.0
python-multipart>=0.0.5
Pillow>=9.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0
pydantic>=1.10.0,<2.0.0
